moving funcs into folder
